+++ app/template/static/js/play.js
@@ -79,7 +79,9 @@ var playFnc = {
         })
 
         if (!actInfo && $(obj).hasClass("att-selected-item")) {
-            var btn = getLinkObj(panelObj, "btn", "chart", "operate", $(obj).attr("data-target"), $(obj).attr("data-type"), "save", "", {"post":"1"});
+            var btnName = "save";
+            if($(obj).attr("data-button-label") && $(obj).attr("data-button-label") != "") btnName = $(obj).attr("data-button-label");
+            var btn = getLinkObj(panelObj, "btn", "chart", "operate", $(obj).attr("data-target"), $(obj).attr("data-type"), btnName, "", {"post":"1"});
             $(act).append(btn);
         }